split checks into functions
